backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
uvloop==0.19.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.15.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
redis[hiredis]==5.0.4
apscheduler==3.10.4
psutil==5.9.8
greenlet==3.0.3
python-dotenv==1.0.1
